nueva UI

+++ controllers/AnaliticasController.php
@@ -24,7 +24,7 @@ switch ($action) {
     case 'getTrafico':
         $mikrotik_id = isset($_GET['mikrotik_id']) ? intval($_GET['mikrotik_id']) : 0;
         $interface = isset($_GET['interface']) && !empty($_GET['interface']) ? $_GET['interface'] : null;
-        $horas = isset($_GET['horas']) ? intval($_GET['horas']) : 4;
+        $horas = isset($_GET['horas']) ? floatval($_GET['horas']) : 4;
         
         if ($mikrotik_id > 0) {
             echo json_encode($dao->getTrafico($mikrotik_id, $interface, $horas));
@@ -35,7 +35,7 @@ switch ($action) {
 
     case 'getPing':
         $mikrotik_id = isset($_GET['mikrotik_id']) ? intval($_GET['mikrotik_id']) : 0;
-        $horas = isset($_GET['horas']) ? intval($_GET['horas']) : 4;
+        $horas = isset($_GET['horas']) ? floatval($_GET['horas']) : 4;
         
         if ($mikrotik_id > 0) {
             echo json_encode($dao->getPing($mikrotik_id, $horas));
@@ -46,7 +46,7 @@ switch ($action) {
 
     case 'getPingEquipo':
         $equipo_id = isset($_GET['equipo_id']) ? intval($_GET['equipo_id']) : 0;
-        $horas = isset($_GET['horas']) ? intval($_GET['horas']) : 4;
+        $horas = isset($_GET['horas']) ? floatval($_GET['horas']) : 4;
         
         if ($equipo_id > 0) {
             echo json_encode($dao->getPingEquipo($equipo_id, $horas));
@@ -57,7 +57,7 @@ switch ($action) {
 
     case 'getRecursos':
         $mikrotik_id = isset($_GET['mikrotik_id']) ? intval($_GET['mikrotik_id']) : 0;
-        $horas = isset($_GET['horas']) ? intval($_GET['horas']) : 4;
+        $horas = isset($_GET['horas']) ? floatval($_GET['horas']) : 4;
         
         if ($mikrotik_id > 0) {
             echo json_encode($dao->getRecursos($mikrotik_id, $horas));
@@ -67,7 +67,7 @@ switch ($action) {
         break;
 
     case 'getTopCaidas':
-        $horas = isset($_GET['horas']) ? intval($_GET['horas']) : 4;
+        $horas = isset($_GET['horas']) ? floatval($_GET['horas']) : 4;
         echo json_encode($dao->getTopCaidas($horas));
         break;
